Fixed sphere occluding a cone
Fixed sphere occluding a cone even when a cone is in front of the sphere. The ray direction inside the cone::hit() was normalized but not inside sphere::hit(). Therefore the t for the intersection point of the cone was always bigger(since normalized) so it always failed the depth test.

Also ported the cone.h to TheNextWeek. Texture mapping is questionable. The Axis Aligned Bounding box is not optimized. Should be easy enough to port it to TheRestOfYourLife.